Firefox fix.
Firefox actually has an object called  window.DeviceOrientationEvent so this test returns true. However, the event is never called. The fix: add another flag to see if the device orientation event is called. On a mobile device it seems to be called immediately, so a 100ms timeout is enough to check if this is working. If the test fails, it is reenabled with mouse move events and works on Firefox.